requirements.txt
# ===================================================================
# FILE: requirements.txt (FINAL – Phase 3: Intent + NER + Semantic Search)
# Tested pada Python 3.9.13
# ===================================================================

# ===================================================================
# FASTAPI & WEB FRAMEWORK
# ===================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
jinja2==3.1.2
python-multipart==0.0.6
starlette==0.27.0

# ===================================================================
# Pydantic & Settings
# ===================================================================
pydantic==2.5.0
pydantic-settings==2.1.0

# ===================================================================
# MACHINE LEARNING & NLP CORE
# ===================================================================
torch==2.1.0    
transformers==4.35.0
tokenizers==0.14.1

# ===================================================================
# NER MODEL (IndoBERT + BiLSTM + CRF)
# ===================================================================
# ⚠️ TorchCRF dari PyPI (1.1.0) tidak mendukung PyTorch >=2.1 → gunakan versi terbaru dari GitHub
git+https://github.com/kmkurn/pytorch-crf.git  # ✅ Fix batch_first error

# ===================================================================
# SEMANTIC SEARCH
# ===================================================================
faiss-cpu==1.7.4
sentence-transformers==2.2.2
rank-bm25==0.2.2

# ===================================================================
# NLP UTILITIES
# ===================================================================
nltk==3.8.1

# ===================================================================
# DATA SCIENCE & UTILITIES
# ===================================================================
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
scipy==1.11.3

# ===================================================================
# LOGGING & MONITORING
# ===================================================================
loguru==0.7.2
python-json-logger==2.0.7
structlog==23.1.0

# ===================================================================
# SECURITY & AUTHENTICATION
# ===================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# ===================================================================
# CACHING & DATABASE
# ===================================================================
redis==5.0.1
aioredis==2.0.1
hiredis==2.2.3

# ===================================================================
# WEB & HTTP CLIENT
# ===================================================================
httpx==0.25.0
requests==2.31.0
urllib3==2.0.7

# ===================================================================
# DEVELOPMENT & TESTING
# ===================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# ===================================================================
# CODE QUALITY & FORMATTING
# ===================================================================
black==23.9.1
isort==5.12.0
flake8==6.1.0
mypy==1.6.1

# ===================================================================
# DOCUMENTATION
# ===================================================================
mkdocs==1.5.3
mkdocs-material==9.4.6

# ===================================================================
# PERFORMANCE & OPTIMIZATION
# ===================================================================
cchardet==2.1.7
orjson==3.9.7

# ===================================================================
# FILE HANDLING & PROCESSING
# ===================================================================
openpyxl==3.1.2
pillow==10.0.1
python-magic==0.4.27

# ===================================================================
# ASYNC & CONCURRENCY
# ===================================================================
asyncio-mqtt==0.16.1
aiofiles==23.2.1
aiodns==3.1.1
